code style suggestions by @WopsS
