+++ WolvenKit.App/Helpers/DiscordHelper.cs
@@ -6,25 +6,19 @@ namespace WolvenKit.Functionality.WKitGlobal.Helpers
 {
     public static class DiscordHelper
     {
-        #region Fields
-
-        public static DiscordRPC.DiscordRpcClient client;
+        public static DiscordRPC.DiscordRpcClient Client;
         public static string DiscordAppID = "897813293480169532";
         public static bool DiscordRPCEnabled = true;
         public static bool DiscordRPCInitizialized = false;
-
-        #endregion Fields
-
-        #region Methods
 
         public static void InitializeDiscordRPC()
         {
             if (DiscordRPCEnabled == true && DiscordRPCInitizialized == false)
             {
-                client = new DiscordRPC.DiscordRpcClient(DiscordAppID);
-                client.OnReady += (sender, e) => Console.WriteLine("Received Ready from user {0}", e.User.Username);
-                client.OnPresenceUpdate += (sender, e) => Console.WriteLine("Received Update! {0}", e.Presence);
-                client.Initialize();
+                Client = new DiscordRPC.DiscordRpcClient(DiscordAppID);
+                Client.OnReady += (sender, e) => Console.WriteLine("Received Ready from user {0}", e.User.Username);
+                Client.OnPresenceUpdate += (sender, e) => Console.WriteLine("Received Update! {0}", e.Presence);
+                Client.Initialize();
                 DiscordRPCInitizialized = true;
             }
         }
@@ -35,21 +29,19 @@ namespace WolvenKit.Functionality.WKitGlobal.Helpers
             {
                 try
                 {
-                    if (client != null)
+                    if (Client != null)
                     {
-                        client.SetPresence(new DiscordRPC.RichPresence()
+                        Client.SetPresence(new DiscordRPC.RichPresence()
                         {
                             Details = details,
                             Timestamps = new DiscordRPC.Timestamps() { Start = DateTime.UtcNow },
                             Assets = new DiscordRPC.Assets() { LargeImageKey = "bigwk", LargeImageText = "WolvenKit", }
                         });
-                        client.Invoke();
+                        Client.Invoke();
                     }
                 }
-                catch (Exception ex){Locator.Current.GetService<ILoggerService>().Error(ex);}
+                catch (Exception ex) { Locator.Current.GetService<ILoggerService>().Error(ex); }
             }
         }
-
-        #endregion Methods
     }
 }
