--- WolvenKit.App/Helpers/GlobalHelper.cs
@@ -1,10 +1,0 @@
-namespace WolvenKit.Functionality.WKitGlobal.Helpers
-{
-    public enum EAnimals
-    {
-        Cat,
-        Dog,
-        Horse,
-        Snail
-    }
-}
