--- WolvenKit.App/Helpers/ImportExport_Helpers.cs
@@ -1,38 +1,0 @@
-namespace WolvenKit.Functionality.Helpers
-{
-    public static class ImportExport_Helpers
-    {
-
-
-
-
-
-    }
-
-
-
-
-    public enum MeshExportOptions
-    {
-        WithPlaceHolderRig,
-        WithoutRig,
-        MultiMeshWithRig,
-        MeshWithRig,
-        MultiMeshWithoutRigh,
-        MeshWithMaterials
-    }
-
-
-
-    public enum WemExportOptions
-    {
-        Wav,
-        MP3,
-    }
-
-
-
-
-
-
-}
