+++ WolvenKit.App/ViewModels/HomePage/Pages/WelcomePageViewModel.cs
@@ -58,18 +58,14 @@ namespace WolvenKit.ViewModels.Shared
             UnpinItem = new DelegateCommand<string>(OnUnpinItemExecute);
             OpenInExplorer = new DelegateCommand<string>(OnOpenInExplorerExecute);
 
-            OpenProjectCommand = ReactiveCommand.Create<string>(s =>
-            {
-                _mainViewModel.OpenProjectCommand.Execute(s).Subscribe();
-            });
-            DeleteProjectCommand = ReactiveCommand.Create<string>(s =>
-            {
-                _mainViewModel.DeleteProjectCommand.Execute(s).Subscribe();
-            });
+            OpenProjectCommand = ReactiveCommand.Create<string>(s => _mainViewModel.OpenProjectCommand.Execute(s).Subscribe());
+            DeleteProjectCommand = ReactiveCommand.Create<string>(s => _mainViewModel.DeleteProjectCommand.Execute(s).Subscribe());
+#pragma warning disable IDE0053 // Doesn't compile with lambda expressions
             NewProjectCommand = ReactiveCommand.Create(() =>
             {
                 _mainViewModel.NewProjectCommand.Execute().Subscribe();
             });
+#pragma warning restore IDE0053 // Doesn't compile with lambda expressions
 
             recentlyUsedItemsService.Items
                 .Connect()
@@ -124,12 +120,7 @@ namespace WolvenKit.ViewModels.Shared
 
         #region Methods
 
-#pragma warning disable AsyncFixer03 // Avoid fire & forget async void methods
-#pragma warning disable AvoidAsyncVoid
-
         private async void OnOpenInExplorerExecute(string parameter)
-#pragma warning restore AsyncFixer03 // Avoid fire & forget async void methods
-#pragma warning restore AvoidAsyncVoid
         {
             if (!File.Exists(parameter))
             {
@@ -208,16 +199,13 @@ namespace WolvenKit.ViewModels.Shared
 
         private void ConvertRecentProjects() // Converts Recent projects for the homepage.
         {
-            DispatcherHelper.RunOnMainThread(() =>
-            {
-                FancyProjects.Clear();
-            });
+            DispatcherHelper.RunOnMainThread(() => FancyProjects.Clear());
 
             var sorted = _recentlyUsedItems.ToList();
             sorted.Sort(delegate (RecentlyUsedItemModel a, RecentlyUsedItemModel b)
             {
                 DateTime ad, bd;
-                if (a.Modified != default(DateTime))
+                if (a.Modified != default)
                 {
                     ad = a.Modified;
                 }
@@ -226,7 +214,7 @@ namespace WolvenKit.ViewModels.Shared
                     ad = a.DateTime;
                 }
 
-                if (b.Modified != default(DateTime))
+                if (b.Modified != default)
                 {
                     bd = b.Modified;
                 }
@@ -258,10 +246,7 @@ namespace WolvenKit.ViewModels.Shared
                     if (!IsThere)
                     { newfi = "pack://application:,,,/Resources/Media/Images/Application/V Male Logo Cropped.png"; }
                     NewItem = new FancyProjectObject(fi.Name, cd, "Cyberpunk 2077", p, newfi);
-                    DispatcherHelper.RunOnMainThread(() =>
-                    {
-                        FancyProjects.Add(NewItem);
-                    });
+                    DispatcherHelper.RunOnMainThread(() => FancyProjects.Add(NewItem));
                 }
                 if (Path.GetExtension(item.Name).TrimStart('.') == EProjectType.w3modproj.ToString())
                 {
@@ -269,10 +254,7 @@ namespace WolvenKit.ViewModels.Shared
                     { newfi = "pack://application:,,,/Resources/Media/Images/Application/tw3proj.png"; }
 
                     NewItem = new FancyProjectObject(n, cd, "The Witcher 3", p, newfi);
-                    DispatcherHelper.RunOnMainThread(() =>
-                    {
-                        FancyProjects.Add(NewItem);
-                    });
+                    DispatcherHelper.RunOnMainThread(() => FancyProjects.Add(NewItem));
                 }
             }
         }
