+++ WolvenKit.App/ViewModels/Shell/ChunkViewModel.cs
@@ -2,19 +2,14 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.ComponentModel;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Reactive.Linq;
-using System.Reflection;
 using System.Runtime.Serialization;
-using System.Text.Json;
 using System.Windows.Forms;
 using System.Windows.Input;
-using DynamicData;
 using DynamicData.Binding;
-using Newtonsoft.Json;
 using ReactiveUI;
 using ReactiveUI.Fody.Helpers;
 using Splat;
@@ -23,21 +18,15 @@ using WolvenKit.Common.Conversion;
 using WolvenKit.Common.Services;
 using WolvenKit.Functionality.Commands;
 using WolvenKit.Functionality.Controllers;
-using WolvenKit.Functionality.Services;
 using WolvenKit.Models;
 using WolvenKit.RED4;
 using WolvenKit.RED4.Archive.Buffer;
 using WolvenKit.RED4.Archive.CR2W;
-using WolvenKit.RED4.Archive.IO;
 using WolvenKit.RED4.CR2W.JSON;
 using WolvenKit.RED4.Types;
 using WolvenKit.ViewModels.Dialogs;
 using WolvenKit.ViewModels.Documents;
 using static WolvenKit.RED4.Types.RedReflection;
-using Mat4 = System.Numerics.Matrix4x4;
-using Quat = System.Numerics.Quaternion;
-using Vec3 = System.Numerics.Vector3;
-using Vec4 = System.Numerics.Vector4;
 
 namespace WolvenKit.ViewModels.Shell
 {
@@ -99,7 +88,7 @@ namespace WolvenKit.ViewModels.Shell
                     CalculateDescriptor();
                     CalculateIsDefault();
 
-                    if (Parent != null)
+                    if (Parent is not null)
                     {
 
                         if (Parent.Data is IRedArray arr)
@@ -113,7 +102,7 @@ namespace WolvenKit.ViewModels.Shell
                             }
                         }
 
-                        if (propertyName != null)
+                        if (propertyName is not null)
                         {
                             var parentType = Parent.PropertyType;
                             var parentData = Parent.Data;
@@ -135,7 +124,7 @@ namespace WolvenKit.ViewModels.Shell
                             else
                             {
                                 var pi = parentData.GetType().GetProperty(propertyName);
-                                if (pi != null)
+                                if (pi is not null)
                                 {
                                     pi.SetValue(parentData, Data);
                                     Tab.File.SetIsDirty(true);
@@ -256,7 +245,7 @@ namespace WolvenKit.ViewModels.Shell
         public void NotifyChain(string property)
         {
             this.RaisePropertyChanged(property);
-            if (Parent != null)
+            if (Parent is not null)
             {
                 Parent.NotifyChain(property);
             }
@@ -293,7 +282,7 @@ namespace WolvenKit.ViewModels.Shell
 
         private readonly RDTDataViewModel _tab;
 
-        public RDTDataViewModel Tab => _tab != null ? _tab : Parent.Tab;
+        public RDTDataViewModel Tab => _tab ?? Parent.Tab;
 
         [Reactive] public IRedType Data { get; set; }
 
@@ -349,7 +338,7 @@ namespace WolvenKit.ViewModels.Shell
             Properties.Clear();
 
             var isreadonly = false;
-            if (Parent != null)
+            if (Parent is not null)
             {
                 isreadonly = Parent.IsReadOnly;
             }
@@ -365,7 +354,7 @@ namespace WolvenKit.ViewModels.Shell
             if (obj is TweakDBID tdb)
             {
                 obj = Locator.Current.GetService<TweakDBService>().GetFlat(tdb);
-                if (obj != null)
+                if (obj is not null)
                 {
                     Properties.Add(new ChunkViewModel(obj, this, "Value")
                     {
@@ -380,7 +369,7 @@ namespace WolvenKit.ViewModels.Shell
                 }
                 isreadonly = true;
                 //var record = Locator.Current.GetService<TweakDBService>().GetRecord(tdb);
-                //if (record != null)
+                //if (record is not null)
                 //{
                 //    Properties.Add(new ChunkViewModel(record, this, "record"));
                 //}
@@ -388,7 +377,7 @@ namespace WolvenKit.ViewModels.Shell
             else if (obj is BaseStringType str)
             {
                 var s = (string)str;
-                if (s != null && s.StartsWith("LocKey#") && ulong.TryParse(s.Substring(7), out var locKey))
+                if (s is not null && s.StartsWith("LocKey#") && ulong.TryParse(s[7..], out var locKey))
                 {
                     obj = Locator.Current.GetService<LocKeyService>().GetEntry(locKey);
                     isreadonly = true;
@@ -474,7 +463,7 @@ namespace WolvenKit.ViewModels.Shell
                         });
                     }
                 }
-                else if (sddb.Data != null)
+                else if (sddb.Data is not null)
                 {
                     var pis = sddb.Data.GetType().GetProperties();
                     foreach (var pi in pis)
@@ -554,7 +543,7 @@ namespace WolvenKit.ViewModels.Shell
                         });
                     }
                 }
-                else if (db.Data != null)
+                else if (db.Data is not null)
                 {
                     var pis = db.Data.GetType().GetProperties();
                     foreach (var pi in pis)
@@ -572,7 +561,7 @@ namespace WolvenKit.ViewModels.Shell
             }
             //else if (Data is TweakXLFile)
             // fallback for non-RTTI data
-            else if (Data != null)
+            else if (Data is not null)
             {
                 if (Data is IBrowsableDictionary ibd)
                 {
@@ -610,7 +599,7 @@ namespace WolvenKit.ViewModels.Shell
                     var pis = Data.GetType().GetProperties();
                     foreach (var pi in pis)
                     {
-                        var value = Data != null ? pi.GetValue(Data) : null;
+                        var value = Data is not null ? pi.GetValue(Data) : null;
                         if (value is IRedType irt)
                         {
                             Properties.Add(new ChunkViewModel(irt, this, pi.Name)
@@ -674,10 +663,10 @@ namespace WolvenKit.ViewModels.Shell
         {
             IsDefault = Data == null;
 
-            if (Parent != null && propertyName != null && Data is not IRedBaseHandle)
+            if (Parent is not null && propertyName is not null && Data is not IRedBaseHandle)
             {
                 var epi = GetPropertyByRedName(Parent.ResolvedPropertyType, propertyName);
-                if (epi != null)
+                if (epi is not null)
                 {
                     //IsDefault = IsDefault(Parent.ResolvedPropertyType, epi, Data);
                     IsDefault = IsDefault(Parent.ResolvedPropertyType, epi, ResolvedData);
@@ -747,18 +736,18 @@ namespace WolvenKit.ViewModels.Shell
             get
             {
                 var type = Data?.GetType() ?? null;
-                if (Parent != null)
+                if (Parent is not null)
                 {
                     var parent = Parent.Data;
                     var parentType = Parent.ResolvedPropertyType;
                     // handles aren't the true parent type of these props, so need to get that
-                    //if (Parent.Data is IRedBaseHandle handle && handle != null)
+                    //if (Parent.Data is IRedBaseHandle handle && handle is not null)
                     //{
                     //    parent = handle.GetValue();
                     //    parentType = handle.GetValue().GetType();
                     //}
                     var propInfo = GetPropertyByRedName(parentType, propertyName) ?? null;
-                    if (propInfo != null)
+                    if (propInfo is not null)
                     {
                         if (type == null || type == propInfo.Type)
                         {
@@ -786,7 +775,7 @@ namespace WolvenKit.ViewModels.Shell
                 if (Data is TweakDBID tdb)
                 {
                     var type = Locator.Current.GetService<TweakDBService>().GetType(tdb);
-                    if (type != null)
+                    if (type is not null)
                     {
                         return type;
                     }
@@ -794,7 +783,7 @@ namespace WolvenKit.ViewModels.Shell
                 if (Data is ITweakXLItem iti)
                 {
                     var type = Locator.Current.GetService<TweakDBService>().GetType(iti.ID);
-                    if (type != null)
+                    if (type is not null)
                     {
                         return type;
                     }
@@ -802,20 +791,20 @@ namespace WolvenKit.ViewModels.Shell
                 if (Data is BaseStringType str)
                 {
                     var s = (string)str;
-                    if (s != null && s.StartsWith("LocKey#") && ulong.TryParse(s.Substring(7), out var _))
+                    if (s is not null && s.StartsWith("LocKey#") && ulong.TryParse(s[7..], out var _))
                     {
                         return typeof(localizationPersistenceOnScreenEntry);
                     }
                 }
-                if (Data is DataBuffer db && db.Data != null)
+                if (Data is DataBuffer db && db.Data is not null)
                 {
                     return db.Data.GetType();
                 }
-                if (Data is SharedDataBuffer sdb && sdb.Data != null)
+                if (Data is SharedDataBuffer sdb && sdb.Data is not null)
                 {
                     return sdb.Data.GetType();
                 }
-                if (Data is SerializationDeferredDataBuffer sddb && sddb.Data != null)
+                if (Data is SerializationDeferredDataBuffer sddb && sddb.Data is not null)
                 {
                     return sddb.Data.GetType();
                 }
@@ -823,7 +812,7 @@ namespace WolvenKit.ViewModels.Shell
                 {
                     return typeof(localizationPersistenceOnScreenEntry);
                 }
-                //if (Data is IBrowsableType ibt && ibt.GetBrowsableType() is var browsableType && browsableType != null)
+                //if (Data is IBrowsableType ibt && ibt.GetBrowsableType() is var browsableType && browsableType is not null)
                 //{
                 //    return browsableType;
                 //}
@@ -835,7 +824,7 @@ namespace WolvenKit.ViewModels.Shell
         {
             get
             {
-                if (PropertyType != null)
+                if (PropertyType is not null)
                 {
                     var redName = GetRedTypeFromCSType(PropertyType, _flags);
                     return redName != "" ? redName : PropertyType.Name;
@@ -844,13 +833,13 @@ namespace WolvenKit.ViewModels.Shell
             }
         }
 
-        public string ResolvedType => ResolvedPropertyType != null ? (GetTypeRedName(ResolvedPropertyType) ?? ResolvedPropertyType.Name) : "";
+        public string ResolvedType => ResolvedPropertyType is not null ? (GetTypeRedName(ResolvedPropertyType) ?? ResolvedPropertyType.Name) : "";
 
         public bool TypesDiffer => PropertyType != ResolvedPropertyType;
 
-        public bool IsInArray => Parent != null && Parent.IsArray;
+        public bool IsInArray => Parent is not null && Parent.IsArray;
 
-        public bool IsArray => PropertyType != null &&
+        public bool IsArray => PropertyType is not null &&
                     (PropertyType.IsAssignableTo(typeof(IRedArray)) ||
                     ResolvedPropertyType.IsAssignableTo(typeof(IList)) ||
                     ResolvedPropertyType.IsAssignableTo(typeof(CR2WList)) ||
@@ -888,7 +877,7 @@ namespace WolvenKit.ViewModels.Shell
                     else if (ResolvedData is BaseStringType str)
                     {
                         var s = (string)str;
-                        if (s != null && s.StartsWith("LocKey#") && ulong.TryParse(s.Substring(7), out var locKey))
+                        if (s is not null && s.StartsWith("LocKey#") && ulong.TryParse(s[7..], out var locKey))
                         {
                             // not actual
                             count += 1;
@@ -913,7 +902,7 @@ namespace WolvenKit.ViewModels.Shell
                         {
                             count += p4.Chunks.Count;
                         }
-                        else if (sddb.Data != null)
+                        else if (sddb.Data is not null)
                         {
                             count += sddb.Data.GetType().GetProperties().Count();
                         }
@@ -952,7 +941,7 @@ namespace WolvenKit.ViewModels.Shell
                             count += 1; // needs refinement?
                         }
                     }
-                    else if (ResolvedData != null)
+                    else if (ResolvedData is not null)
                     {
                         if (Data is IBrowsableDictionary ibd)
                         {
@@ -990,7 +979,7 @@ namespace WolvenKit.ViewModels.Shell
             get
             {
                 var width = 0;
-                if (Parent != null)
+                if (Parent is not null)
                 {
                     //if (Parent.ResolvedData is IRedArray ary)
                     //{
@@ -1080,7 +1069,7 @@ namespace WolvenKit.ViewModels.Shell
                 if (!string.IsNullOrEmpty(value))
                 {
                     Value = value;
-                    if (Value != null && Value.StartsWith("LocKey#") && ulong.TryParse(Value.Substring(7), out var key))
+                    if (Value is not null && Value.StartsWith("LocKey#") && ulong.TryParse(Value[7..], out var key))
                     {
                         Value = "";
                         //    Value = Locator.Current.GetService<LocKeyService>().GetFemaleVariant(key);
@@ -1152,7 +1141,7 @@ namespace WolvenKit.ViewModels.Shell
             else if (PropertyType.IsAssignableTo(typeof(IRedRef)))
             {
                 var value = (IRedRef)Data;
-                Value = value != null && value.DepotPath.GetResolvedText() != "" ? value.DepotPath.GetResolvedText() : "null";
+                Value = value is not null && value.DepotPath.GetResolvedText() != "" ? value.DepotPath.GetResolvedText() : "null";
             }
             else if (Data is IBrowsableType ibt)
             {
@@ -1211,14 +1200,14 @@ namespace WolvenKit.ViewModels.Shell
             else if (Data is BaseStringType str)
             {
                 var s = (string)str;
-                if (s != null && s.StartsWith("LocKey#") && ulong.TryParse(s.Substring(7), out var locKey2))
+                if (s is not null && s.StartsWith("LocKey#") && ulong.TryParse(s[7..], out var locKey2))
                 {
                     Descriptor = locKey2.ToString();
                 }
             }
-            //if (ResolvedData is CMaterialInstance && Parent != null)
+            //if (ResolvedData is CMaterialInstance && Parent is not null)
             //{
-            //    if (Parent.Parent != null && Parent.Parent.Parent != null && Parent.Parent.Data is CMesh mesh)
+            //    if (Parent.Parent is not null && Parent.Parent.Parent is not null && Parent.Parent.Data is CMesh mesh)
             //    {
             //        Descriptor = mesh.MaterialEntries[int.Parse(Name)].Name;
             //    }
@@ -1235,7 +1224,7 @@ namespace WolvenKit.ViewModels.Shell
             {
                 Descriptor = $"{q.I}, {q.J}, {q.K}, {q.R}";
             }
-            if (Data is CMaterialInstance && Parent != null && Tab.File.Cr2wFile.RootChunk is CMesh mesh)
+            if (Data is CMaterialInstance && Parent is not null && Tab.File.Cr2wFile.RootChunk is CMesh mesh)
             {
                 if (mesh.LocalMaterialBuffer.RawData.Data is CR2WList list)
                 {
@@ -1252,7 +1241,7 @@ namespace WolvenKit.ViewModels.Shell
                     }
                 }
             }
-            else if (ResolvedData != null)
+            else if (ResolvedData is not null)
             {
                 if (Data is IBrowsableDictionary ibd)
                 {
@@ -1291,7 +1280,7 @@ namespace WolvenKit.ViewModels.Shell
                 foreach (var propName in propNames)
                 {
                     var prop = GetPropertyByRedName(irc.GetType(), propName);
-                    if (prop != null)
+                    if (prop is not null)
                     {
                         Descriptor = irc.GetProperty(prop.RedName).ToString();
                         return;
@@ -1302,10 +1291,10 @@ namespace WolvenKit.ViewModels.Shell
             {
                 foreach (var propName in propNames)
                 {
-                    if (Data != null)
+                    if (Data is not null)
                     {
                         var prop = Data.GetType().GetProperty(propName);
-                        if (prop != null)
+                        if (prop is not null)
                         {
                             Descriptor = prop.GetValue(Data).ToString();
                             return;
@@ -1392,7 +1381,7 @@ namespace WolvenKit.ViewModels.Shell
         public bool CanBeDroppedOn(ChunkViewModel target) => PropertyType == target.PropertyType;
 
         public ICommand OpenRefCommand { get; private set; }
-        private bool CanOpenRef() => Data is IRedRef r && r.DepotPath != null;
+        private bool CanOpenRef() => Data is IRedRef r && r.DepotPath is not null;
         private void ExecuteOpenRef()
         {
             if (Data is IRedRef r)
@@ -1413,7 +1402,7 @@ namespace WolvenKit.ViewModels.Shell
         }
 
         public ICommand AddRefCommand { get; private set; }
-        private bool CanAddRef() => Data is IRedRef r && r.DepotPath != null;
+        private bool CanAddRef() => Data is IRedRef r && r.DepotPath is not null;
         private void ExecuteAddRef()
         {
             if (Data is IRedRef r)
@@ -1453,7 +1442,7 @@ namespace WolvenKit.ViewModels.Shell
         {
             var app = Locator.Current.GetService<AppViewModel>();
             app.CloseDialogCommand.Execute(null);
-            if (sender != null)
+            if (sender is not null)
             {
                 var vm = sender as CreateClassDialogViewModel;
                 var instance = RedTypeManager.Create(vm.SelectedClass);
@@ -1550,7 +1539,7 @@ namespace WolvenKit.ViewModels.Shell
         }
 
         public ICommand AddItemToCompiledDataCommand { get; private set; }
-        private bool CanAddItemToCompiledData() => ResolvedPropertyType != null && ResolvedPropertyType.IsAssignableTo(typeof(IRedBufferPointer));
+        private bool CanAddItemToCompiledData() => ResolvedPropertyType is not null && ResolvedPropertyType.IsAssignableTo(typeof(IRedBufferPointer));
         private void ExecuteAddItemToCompiledData()
         {
             if (Data == null)
@@ -1589,7 +1578,7 @@ namespace WolvenKit.ViewModels.Shell
         {
             var app = Locator.Current.GetService<AppViewModel>();
             app.CloseDialogCommand.Execute(null);
-            if (sender != null)
+            if (sender is not null)
             {
                 var vm = sender as SelectRedTypeDialogViewModel;
 
@@ -1602,7 +1591,7 @@ namespace WolvenKit.ViewModels.Shell
         {
             var app = Locator.Current.GetService<AppViewModel>();
             app.CloseDialogCommand.Execute(null);
-            if (sender != null)
+            if (sender is not null)
             {
                 var vm = sender as CreateClassDialogViewModel;
                 var instance = RedTypeManager.Create(vm.SelectedClass);
@@ -1614,7 +1603,7 @@ namespace WolvenKit.ViewModels.Shell
         {
             var app = Locator.Current.GetService<AppViewModel>();
             app.CloseDialogCommand.Execute(null);
-            if (sender != null)
+            if (sender is not null)
             {
                 var vm = sender as CreateClassDialogViewModel;
                 var instance = RedTypeManager.Create(vm.SelectedClass);
@@ -1859,7 +1848,7 @@ namespace WolvenKit.ViewModels.Shell
 
                 if (saveFileDialog.ShowDialog() == DialogResult.OK)
                 {
-                    if ((myStream = saveFileDialog.OpenFile()) != null)
+                    if ((myStream = saveFileDialog.OpenFile()) is not null)
                     {
                         var json = RedJsonSerializer.Serialize(irt);
                         if (!string.IsNullOrEmpty(json))
@@ -1889,7 +1878,7 @@ namespace WolvenKit.ViewModels.Shell
 
             if (saveFileDialog.ShowDialog() == DialogResult.OK)
             {
-                if ((myStream = saveFileDialog.OpenFile()) != null)
+                if ((myStream = saveFileDialog.OpenFile()) is not null)
                 {
                     var dto = new RedTypeDto(ResolvedData);
                     var json = RedJsonSerializer.Serialize(dto);
@@ -1906,7 +1895,7 @@ namespace WolvenKit.ViewModels.Shell
         }
 
         public ICommand OpenChunkCommand { get; private set; }
-        private bool CanOpenChunk() => Data is RedBaseClass && Parent != null;
+        private bool CanOpenChunk() => Data is RedBaseClass && Parent is not null;
         private void ExecuteOpenChunk()
         {
             if (Data is RedBaseClass cls)
@@ -2149,7 +2138,7 @@ namespace WolvenKit.ViewModels.Shell
 
         public ICommand PasteChunkCommand { get; private set; }
         private bool CanPasteChunk() => (IsArray || IsInArray)
-            && RDTDataViewModel.CopiedChunk != null
+            && RDTDataViewModel.CopiedChunk is not null
             && (ArraySelfOrParent?.InnerType.IsAssignableTo(RDTDataViewModel.CopiedChunk.GetType()) ?? true);
         private void ExecutePasteChunk()
         {
@@ -2220,7 +2209,7 @@ namespace WolvenKit.ViewModels.Shell
                 }
             }
 
-            if (sourceList != null && destList != null)
+            if (sourceList is not null && destList is not null)
             {
                 int oldIndex = -1, i = 0;
                 foreach (var thing in sourceList)
@@ -2357,7 +2346,7 @@ namespace WolvenKit.ViewModels.Shell
 
             IsExpanded = true;
 
-            if (selectChild != null)
+            if (selectChild is not null)
             {
                 foreach (var prop in Properties)
                 {
