+++ WolvenKit.Core/Extensions/StringExtensions.cs
@@ -11,22 +11,19 @@ namespace WolvenKit.Interfaces.Extensions
     {
         public static string FirstCharToLower(this string input)
         {
-            switch (input)
+            return input switch
             {
-                case null:
-                    throw new ArgumentNullException(nameof(input));
-                case "":
-                    throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input));
-                default:
-                    return input.First().ToString().ToLower() + input.Substring(1);
-            }
+                null => throw new ArgumentNullException(nameof(input)),
+                "" => throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input)),
+                _ => input.First().ToString().ToLower() + input[1..],
+            };
         }
 
         public static string FirstCharToUpper(this string input) => input switch
         {
             null => throw new ArgumentNullException(nameof(input)),
             "" => throw new ArgumentException($"{nameof(input)} cannot be empty", nameof(input)),
-            _ => input.First().ToString().ToUpper() + input.Substring(1)
+            _ => input.First().ToString().ToUpper() + input[1..]
         };
 
         // https://stackoverflow.com/a/3695190
@@ -53,7 +50,7 @@ namespace WolvenKit.Interfaces.Extensions
 
         public static (string, bool, EProjectFolders) GetModRelativePath(this string fullpath, string activeModFileDirectory)
         {
-            var relativePath = fullpath.Substring(activeModFileDirectory.Length + 1);
+            var relativePath = fullpath[(activeModFileDirectory.Length + 1)..];
             bool isDLC;
             var projectfolder = EProjectFolders.Cooked;
 
@@ -70,26 +67,26 @@ namespace WolvenKit.Interfaces.Extensions
                 throw new NotImplementedException();
             }
 
-            relativePath = relativePath.Substring(4);
+            relativePath = relativePath[4..];
 
             if (relativePath.StartsWith(EProjectFolders.Cooked.ToString()))
             {
-                relativePath = relativePath.Substring(EProjectFolders.Cooked.ToString().Length + 1);
+                relativePath = relativePath[(EProjectFolders.Cooked.ToString().Length + 1)..];
                 projectfolder = EProjectFolders.Cooked;
             }
 
             if (relativePath.StartsWith(EProjectFolders.Uncooked.ToString()))
             {
-                relativePath = relativePath.Substring(EProjectFolders.Uncooked.ToString().Length + 1);
+                relativePath = relativePath[(EProjectFolders.Uncooked.ToString().Length + 1)..];
                 projectfolder = EProjectFolders.Uncooked;
             }
             else if (relativePath.StartsWith(EArchiveType.SoundCache.ToString()))
             {
-                relativePath = relativePath.Substring(EArchiveType.SoundCache.ToString().Length + 1);
+                relativePath = relativePath[(EArchiveType.SoundCache.ToString().Length + 1)..];
             }
             else if (relativePath.StartsWith(EArchiveType.Speech.ToString()))
             {
-                relativePath = relativePath.Substring(EArchiveType.Speech.ToString().Length + 1);
+                relativePath = relativePath[(EArchiveType.Speech.ToString().Length + 1)..];
             }
 
             return (relativePath, isDLC, projectfolder);
@@ -117,7 +114,7 @@ namespace WolvenKit.Interfaces.Extensions
             var result = target;
             while (result.StartsWith(trimString))
             {
-                result = result.Substring(trimString.Length);
+                result = result[trimString.Length..];
             }
 
             return result;
