--- WolvenKit/Views/HomePage/Pages/DebugPageView.xaml
@@ -1,12 +1,0 @@
-<reactiveUi:ReactiveUserControl
-    x:TypeArguments="pages:DebugPageViewModel"
- x:Class="WolvenKit.Views.HomePage.Pages.DebugPageView"
- xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
- xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
- xmlns:reactiveUi="http://reactiveui.net"
- xmlns:hc="https://handyorg.github.io/handycontrol"
- xmlns:svgc="http://sharpvectors.codeplex.com/svgc/"
- xmlns:pages="clr-namespace:WolvenKit.ViewModels.HomePage.Pages;assembly=WolvenKit.App">
-
-    <Grid />
-</reactiveUi:ReactiveUserControl>
