--- WolvenKit/Views/HomePage/Pages/DebugPageView.xaml.cs
@@ -1,21 +1,0 @@
-using ReactiveUI;
-using Splat;
-using WolvenKit.ViewModels.HomePage.Pages;
-
-namespace WolvenKit.Views.HomePage.Pages
-{
-    public partial class DebugPageView : ReactiveUserControl<DebugPageViewModel>
-    {
-        #region Constructors
-
-        public DebugPageView()
-        {
-            InitializeComponent();
-
-            ViewModel = Locator.Current.GetService<DebugPageViewModel>();
-            DataContext = ViewModel;
-        }
-
-        #endregion Constructors
-    }
-}
